==> 04_Flux_QC/03_precip_filter.Rmd <==
---
title: "03) Precipitation filtering (TUCC)"
author: "Stenka Vulova"
date: "27 Oct. 2021"
output:
  pdf_document: default
  html_document: default
---

**2021 update:** "EddyPro output data at 30-min resolution with quality flag 2 were excluded (Foken, 2016)as well as data during and **4 h after precipitation events based upon DWD precipitation data**." (Vulova et al. 2021)

I am importing data which is already quality-checked/ filtered using Freddypro. 

# Libraries 

```{r libraries, include = FALSE }

library(sp)
library(raster)
library(rgdal)
library(lubridate)
library(maptools)
library(ggplot2)
library(plyr)
library(dplyr)
library(grid)
library(spatstat)
library(sf)
library(readr)
library(RColorBrewer)
library(reshape2)
library(scales)

library(tictoc) # benchmarking 
library(spatialEco) # for weighted_footprint function 

library(openair)
library(water)
library(stats)

library(tidyverse)  # separate function
library(xts)        # time series
#library(TigR)       # time series apply.hourly function
library(MeTo)       # VP function
#library(FREddyPro)  # cleanFluxes with quality flag - remove outlier and clean SD
library(suncalc)    # getSunlightTimes

```

The older package is no longer in the CRAN. 
https://cran.r-project.org/src/contrib/Archive/FREddyPro/ 

https://support.rstudio.com/hc/en-us/articles/219949047-Installing-older-versions-of-packages 

```{r Freddypro library }

# https://cran.r-project.org/src/contrib/Archive/FREddyPro/FREddyPro_1.0.tar.gz

#packageurl <- "https://cran.r-project.org/src/contrib/Archive/FREddyPro/FREddyPro_1.0.tar.gz"
#install.packages(packageurl, repos=NULL, type="source")

library(FREddyPro)
  
```

# Data import 

I previously created LE data filtered by Freddypro.

```{r data import }

load("/Users/stenkavulova/Dropbox/Stenka_UWI/Topic_3/04_PROCESSED_DATA/20211026_QC/TUCC/02_Freddypro_QC/TUCC_cleaned.RData")

```

# Create a factor to mask precipitation using DWD station

## Prepare TUCC_cleaned

I believe that for the code to work, there has to be half-hourly data for all cases (no missing data). 

```{r prepare TUCC_cleaned }

# Prepare TUCC_cleaned 
# Originally 46127 rows in the dataframe 

min(TUCC_cleaned$timestamp)
max(TUCC_cleaned$timestamp)

# https://stackoverflow.com/questions/16787038/insert-rows-for-missing-dates-times

ts = seq.POSIXt(min(TUCC_cleaned$timestamp), as.POSIXlt(max(TUCC_cleaned$timestamp)), by="30 mins")

ts = format.POSIXct(ts,"%Y-%m-%d %H:%M:%S")

df <- data.frame(timestamp=ts)

df$timestamp = as.POSIXct(df$timestamp, format = "%Y-%m-%d %H:%M:%S", tz = "UTC")

nrow(df)

TUCC_cleaned2 = dplyr::full_join(TUCC_cleaned, df)

nrow(TUCC_cleaned) 

# Remove duplicate timestamp rows
TUCC_cleaned2 = TUCC_cleaned2 %>% distinct(timestamp, .keep_all = TRUE)

nrow(TUCC_cleaned2)

identical(TUCC_cleaned$timestamp, df$timestamp) # TRUE 

```
In this case, there was no missing half-hourly data (already took care of that in previous codes).

## Make precip_factor

Alby already made this for the DWD data. 

Alby: "The raining window calculation, replacing the first for the second function works, I tested with your data.  
`row_timestamp <- data.frame(row=row_number(EC.DWD_clean$timestamp))`
`row_timestamp <- data.frame(row=1:length(TUCC_cleaned$timestamp))` "  

Alby: "I renamed the 0h, because was the same to the variable name. So test if it will correct, below in red.
It is important to do in the sequence from the higher to 0." 

```{r create precip factor }

#TUCC_cleaned2$precip.no.yes

window.prec.9h <- arrange(data.frame(row=unique(c(which(TUCC_cleaned2$precip.no.yes==1) + 9))), row)
window.prec.9h$w9h <- 9

window.prec.8h <- arrange(data.frame(row=unique(c(which(TUCC_cleaned2$precip.no.yes==1) + 8))), row)
window.prec.8h$w8h <- 8

window.prec.7h <- arrange(data.frame(row=unique(c(which(TUCC_cleaned2$precip.no.yes==1) + 7))), row)
window.prec.7h$w7h <- 7

window.prec.6h <- arrange(data.frame(row=unique(c(which(TUCC_cleaned2$precip.no.yes==1) + 6))), row)
window.prec.6h$w6h <- 6

window.prec.5h <- arrange(data.frame(row=unique(c(which(TUCC_cleaned2$precip.no.yes==1) + 5))), row)
window.prec.5h$w5h <- 5

window.prec.4h <- arrange(data.frame(row=unique(c(which(TUCC_cleaned2$precip.no.yes==1) + 4))), row)
window.prec.4h$w4h <- 4

window.prec.3h <- arrange(data.frame(row=unique(c(which(TUCC_cleaned2$precip.no.yes==1) + 3))), row)
window.prec.3h$w3h <- 3

window.prec.2h <- arrange(data.frame(row=unique(c(which(TUCC_cleaned2$precip.no.yes==1) + 2))), row)
window.prec.2h$w2h <- 2

window.prec.1h <- arrange(data.frame(row=unique(c(which(TUCC_cleaned2$precip.no.yes==1) + 1))), row)
window.prec.1h$w1h <- 1

window.prec.0h <- arrange(data.frame(row=unique(c(which(TUCC_cleaned2$precip.no.yes==1) ))), row)
window.prec.0h$hw <- 0

# replaced this line (the df row numbering previously caused an issue)
row_timestamp <- data.frame(row=1:length(TUCC_cleaned2$timestamp))
row_timestamp$timestamp <- TUCC_cleaned2$timestamp

row_timestamp <- left_join(row_timestamp, window.prec.0h, by="row")
row_timestamp <- left_join(row_timestamp, window.prec.1h, by="row")
row_timestamp <- left_join(row_timestamp, window.prec.2h, by="row")
row_timestamp <- left_join(row_timestamp, window.prec.3h, by="row")
row_timestamp <- left_join(row_timestamp, window.prec.4h, by="row")
row_timestamp <- left_join(row_timestamp, window.prec.5h, by="row")
row_timestamp <- left_join(row_timestamp, window.prec.6h, by="row")
row_timestamp <- left_join(row_timestamp, window.prec.7h, by="row")
row_timestamp <- left_join(row_timestamp, window.prec.8h, by="row")
row_timestamp <- left_join(row_timestamp, window.prec.9h, by="row")

row_timestamp$prec.window <- "5 hours or more"
row_timestamp$prec.window[row_timestamp$w9h==9] <- "4.5 hours"
row_timestamp$prec.window[row_timestamp$w8h==8] <- "4 hours"
row_timestamp$prec.window[row_timestamp$w7h==7] <- "3.5 hours"
row_timestamp$prec.window[row_timestamp$w6h==6] <- "3 hours"
row_timestamp$prec.window[row_timestamp$w5h==5] <- "2.5 hours"
row_timestamp$prec.window[row_timestamp$w4h==4] <- "2 hours" # 2 hours 
row_timestamp$prec.window[row_timestamp$w3h==3] <- "1.5 hours" # 1.5 hours
row_timestamp$prec.window[row_timestamp$w2h==2] <- "1 hour" # 1 hour
row_timestamp$prec.window[row_timestamp$w1h==1] <- "0.5 hours" # 0.5 hours 
row_timestamp$prec.window[row_timestamp$hw==0] <- "0 hour"

TUCC_cleaned2$prec.window <- row_timestamp$prec.window

```

Let's look at the data. 

```{r precipitation data}

precip_check = 
  TUCC_cleaned2 %>%
  dplyr::select(timestamp, precip.no.yes, prec.window)

# save the data 

TUCC_QC_precip = TUCC_cleaned2

#save(TUCC_QC_precip, file = "/Users/stenkavulova/Dropbox/Stenka_UWI/Topic_3/04_PROCESSED_DATA/20211026_QC/TUCC/03_precip_filter/TUCC_QC_precip.RData")

#write.csv(TUCC_QC_precip, file = "/Users/stenkavulova/Dropbox/Stenka_UWI/Topic_3/04_PROCESSED_DATA/20211026_QC/TUCC/03_precip_filter/TUCC_QC_precip.csv", row.names = FALSE )

```
